Silence warnings when verified they are OK

+++ FreeRTOS-Plus/Source/FreeRTOS-Plus-TCP/FreeRTOS_Sockets.c
@@ -3415,7 +3415,7 @@ void vSocketWakeUpUser( FreeRTOS_Socket_t *pxSocket )
 						( int ) pxSocket->u.xTCP.usBacklog);
 					/* These should never evaluate to false since the buffers are both shorter than 5-6 characters (<=65535) */
 					configASSERT( copied_len >= 0 );
-					configASSERT( copied_len < sizeof( ucChildText ) );
+					configASSERT( copied_len < (int32_t) sizeof( ucChildText ) );
 				}
 				FreeRTOS_printf( ( "TCP %5d %-16lxip:%5d %d/%d %-13.13s %6lu %6u%s\n",
 					pxSocket->usLocalPort,		/* Local port on this machine */
